requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
youtube-transcript-api>=1.0.0
apscheduler>=3.10.0
structlog>=24.0.0
